requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
mailersend==0.5.8
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyodbc==5.2.0
psycopg2-binary==2.9.9
pyparsing==3.2.1
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
resend==2.10.0
rsa==4.9
schedule==1.2.0
sendgrid==6.11.0
soupsieve==2.6
SQLAlchemy==2.0.37
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0